// src/commons/types/generated/types.s.ts
export type Maybe<T> = T | null;
export type Exact<T extends { [key: string]: unknown }> = {
  [K in keyof T]: T[K];
};
export type MakeOptional<T, K extends keyof T> = Omit<T, K> &
  { [SubKey in K]?: Maybe<T[SubKey]> };
export type MakeMaybe<T, K extends keyof T> = Omit<T, K> &
  { [SubKey in K]: Maybe<T[SubKey]> };
/** All built-in and custom scalars, mapped to their actual values */
export type Scalars = {
  ID: string;
  String: string;
  Boolean: boolean;
  Int: number;
  Float: number;
  /** A date-time string at UTC, such as 2019-12-03T09:54:33Z, compliant with the date-time format. */
  DateTime: any;
};

export type IBoard = {
  __typename?: "Board";
  _id: Scalars["ID"];
  writer?: Maybe<Scalars["String"]>;
  title: Scalars["String"];
  contents: Scalars["String"];
  youtubeUrl: Scalars["String"];
  likeCount: Scalars["Int"];
  dislikeCount: Scalars["Int"];
  boardAddress?: Maybe<IBoardAddress>;
  user?: Maybe<IUser>;
  createdAt: Scalars["DateTime"];
  updatedAt: Scalars["DateTime"];
  deletedAt?: Maybe<Scalars["DateTime"]>;
};

export type IBoardAddress = {
  __typename?: "BoardAddress";
  _id: Scalars["ID"];
  zipcode: Scalars["String"];
  address: Scalars["String"];
  addressDetail: Scalars["String"];
  createdAt: Scalars["DateTime"];
  updatedAt: Scalars["DateTime"];
  deletedAt?: Maybe<Scalars["DateTime"]>;
};

export type IBoardComment = {
  __typename?: "BoardComment";
  _id: Scalars["ID"];
  writer?: Maybe<Scalars["String"]>;
  contents: Scalars["String"];
  rating: Scalars["Float"];
  user?: Maybe<IUser>;
  createdAt: Scalars["DateTime"];
  updatedAt: Scalars["DateTime"];
  deletedAt?: Maybe<Scalars["DateTime"]>;
};
// ...
export type IMutation = {
  __typename?: "Mutation";
  createBoard: IBoard;
  deleteBoards: Array<Scalars["ID"]>;
  dislikeBoard: Scalars["Int"];
  likeBoard: Scalars["Int"];
  updateBoard: IBoard;
  createBoardComment: IBoardComment;
  deleteBoardComment: Scalars["ID"];
  updateBoardComment: IBoardComment;
  createPointTransactionOfBuyingAndSelling: IUseditem;
  createPointTransactionOfLoading: IPointTransaction;
  createUseditem: IUseditem;
  updateUseditem: IUseditem;
  toggleUseditemPick: Scalars["Int"];
  createUseditemQuestion: IUseditemQuestion;
  deleteUseditemQuestion: Scalars["ID"];
  updateUseditemQuestion: IUseditemQuestion;
  createUseditemQuestionAnswer: IUseditemQuestionAnswer;
  deleteUseditemQuestionAnswer: Scalars["String"];
  updateUseditemQuestionAnswer: IUseditemQuestionAnswer;
  createUser: IUser;
  logoutUser: Scalars["Boolean"];
  resetUserPassword: Scalars["Boolean"];
  updateUser: Scalars["Boolean"];
};

export type IMutationCreateBoardArgs = {
  createBoardInput: ICreateBoardInput;
};

export type IMutationDeleteBoardsArgs = {
  boardIds: Array<Scalars["ID"]>;
};

export type IMutationDislikeBoardArgs = {
  boardId: Scalars["ID"];
};

export type IMutationLikeBoardArgs = {
  boardId: Scalars["ID"];
};

export type IMutationUpdateBoardArgs = {
  updateBoardInput: IUpdateBoardInput;
  password?: Maybe<Scalars["String"]>;
  boardId: Scalars["ID"];
};
// ...
export type IPointTransaction = {
  __typename?: "PointTransaction";
  _id: Scalars["ID"];
  impUid: Scalars["ID"];
  amount: Scalars["Int"];
  balance: Scalars["Int"];
  status: Scalars["String"];
  statusDetail: Scalars["String"];
  user?: Maybe<IUser>;
  createdAt: Scalars["DateTime"];
  updatedAt: Scalars["DateTime"];
  deletedAt?: Maybe<Scalars["DateTime"]>;
};
// ...
export type IUpdateBoardInput = {
  title?: Maybe<Scalars["String"]>;
  contents?: Maybe<Scalars["String"]>;
  youtubeUrl?: Maybe<Scalars["String"]>;
};

export type IUpdateUseditemInput = {
  name: Scalars["String"];
  remarks: Scalars["String"];
  contents: Scalars["String"];
  price: Scalars["Int"];
  tags?: Maybe<Array<Scalars["String"]>>;
};

export type IUpdateUseditemQuestionAnswerInput = {
  contents: Scalars["String"];
};

export type IUpdateUseditemQuestionInput = {
  contents: Scalars["String"];
};

export type IUseditem = {
  __typename?: "Useditem";
  _id: Scalars["ID"];
  name: Scalars["String"];
  remarks: Scalars["String"];
  contents: Scalars["String"];
  price: Scalars["Int"];
  tags: Array<Scalars["String"]>;
  useditemAddress?: Maybe<IUseditemAddress>;
  buyer: IUser;
  seller: IUser;
  soldAt?: Maybe<Scalars["DateTime"]>;
  createdAt: Scalars["DateTime"];
  updatedAt: Scalars["DateTime"];
  deletedAt?: Maybe<Scalars["DateTime"]>;
};

export type IUseditemAddress = {
  __typename?: "UseditemAddress";
  id: Scalars["ID"];
  zipcode: Scalars["String"];
  address: Scalars["String"];
  addressDetail: Scalars["String"];
  lat: Scalars["Int"];
  lng: Scalars["Int"];
};

export type IUseditemQuestion = {
  __typename?: "UseditemQuestion";
  _id: Scalars["ID"];
  contents: Scalars["String"];
  useditem: IUseditem;
  user: IUser;
  createdAt: Scalars["DateTime"];
  updatedAt: Scalars["DateTime"];
  deletedAt?: Maybe<Scalars["DateTime"]>;
};

export type IUseditemQuestionAnswer = {
  __typename?: "UseditemQuestionAnswer";
  _id: Scalars["ID"];
  contents: Scalars["String"];
  useditemQuestion: IUseditemQuestion;
  user: IUser;
  createdAt: Scalars["DateTime"];
  updatedAt: Scalars["DateTime"];
  deletedAt?: Maybe<Scalars["DateTime"]>;
};

export type IUser = {
  __typename?: "User";
  _id: Scalars["ID"];
  email: Scalars["String"];
  name: Scalars["String"];
  userPoint?: Maybe<IUserPoint>;
  createdAt: Scalars["DateTime"];
  updatedAt: Scalars["DateTime"];
  deletedAt?: Maybe<Scalars["DateTime"]>;
};

export type IUserPoint = {
  __typename?: "UserPoint";
  _id: Scalars["ID"];
  amount: Scalars["Int"];
  user: IUser;
  createdAt: Scalars["DateTime"];
  updatedAt: Scalars["DateTime"];
  deletedAt?: Maybe<Scalars["DateTime"]>;
};